Tried SA + Changes in simulation function
Simulation function became far too big with the two algorithms, so made
functions for each algorithm. Main variables are in the __name__ ==
__main__.
SA algorithm does NOT work correctly. It will run, but values are not
correct, having difficulties with copying the land and houses. Tried to
make a function called createNewLand which in theory makes a copy, but
not sure if this will work.